python_scripts/requirements.txt
joblib==1.3.1
numpy==1.25.1
pandas==2.0.3
PyMySQL==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
scikit-surprise==1.1.3
scipy==1.11.1
six==1.16.0
surprise==0.1
tzdata==2023.3
